+++ .gitignore
@@ -1,2 +1,7 @@
 *.pyc 
 .DS_Store
+*.egg-info/
+dist/
+build/
+*.egg
+__pycache__/
