+++ opencv_document_scanner/pyimagesearch/__init__.py
@@ -1,0 +1,1 @@
+from . import *  # noqa: F403
